e2b-dev/agent-protocol/tests/test_sealed_auction.py
```python
import unittest
from e2b-dev.agent-protocol.sealed_auction import SealedAuction
from e2b-dev.agent-protocol.agent import Agent
from e2b-dev.agent-protocol.queue import Queue
from e2b-dev.agent-protocol.message import Message

class TestSealedAuction(unittest.TestCase):

    def setUp(self):
        self.auction = SealedAuction()
        self.agent1 = Agent("Agent1")
        self.agent2 = Agent("Agent2")
        self.queue = Queue()

    def test_bid(self):
        self.auction.bid(self.agent1, 100)
        self.assertEqual(self.auction.bids[self.agent1], 100)

    def test_resolve(self):
        self.auction.bid(self.agent1, 100)
        self.auction.bid(self.agent2, 200)
        winner, amount = self.auction.resolve()
        self.assertEqual(winner, self.agent2)
        self.assertEqual(amount, 200)

    def test_message_passing(self):
        message = Message(self.agent1, self.agent2, "Bid", 100)
        self.agent1.send_message(message)
        received_message = self.agent2.receive_message()
        self.assertEqual(received_message, message)

    def test_queue_work(self):
        task = "Task1"
        self.queue.add_task(task)
        self.assertEqual(self.queue.get_task(), task)

if __name__ == '__main__':
    unittest.main()
```